Extract ExternalLink and PageLayout into separate folders
