+++ components/ExternalLink/ExternalLink.js
@@ -1,0 +1,10 @@
+import styles from "./ExternalLink.module.scss";
+export default function ExternalLink({ url, children }) {
+  return (
+    <>
+      <a href={url} target="_blank" className={styles.externalLink}>
+        {children}
+      </a>
+    </>
+  );
+}
